Changed to NGSi v2, removed stamp and more...
Fiware-Object-Converter is now more integrated into this Project.
The CBSubscriber and CBPublisher are completely rewritten. Obj2Ros now loads the needed Messages directly (currently unefficient). The files TopicHandler and RequestHandler are more or less the same. They beoth need to be refactored.

Due to the new Converter and NGSI-API, the firossstamp was no longer needed and is now removed.